+++ lib/legion/extensions/mesh.rb
@@ -7,6 +7,7 @@ require 'legion/extensions/mesh/helpers/preference_profile'
 require 'legion/extensions/mesh/helpers/pending_requests'
 require 'legion/extensions/mesh/helpers/delegation'
 require 'legion/extensions/mesh/helpers/peer_verify'
+require 'legion/extensions/mesh/helpers/peer_table'
 require 'legion/extensions/mesh/runners/mesh'
 require 'legion/extensions/mesh/runners/preferences'
 require 'legion/extensions/mesh/runners/delegation'
@@ -16,7 +17,10 @@ if Legion.const_defined?(:Transport, false)
   require 'legion/extensions/mesh/transport/messages/preference_query'
   require 'legion/extensions/mesh/transport/messages/preference_response'
   require 'legion/extensions/mesh/transport/messages/mesh_departure'
+  require 'legion/extensions/mesh/transport/messages/gossip'
+  require 'legion/extensions/mesh/transport/messages/mesh_conflict'
   require 'legion/extensions/mesh/transport/queues/preference'
+  require 'legion/extensions/mesh/transport/queues/gossip'
 end
 
 module Legion